digistar-indent-line-function: begin rewrite, allowing tabs
The previous implementation of digistar-indent-line-function proved to
be severely broken in a couple of ways.  It didn't play nice with
electric-indent, which is enabled by default since Emacs 24.4, and I
decided to give in and just use tabs.  The digistar 5 script editor
itself uses tabs so there is no getting away from them, and furthermore,
Digistar 5 uses a pathological tab-width of 7, so files indented with
spaces to the usual 8 end up very messy when edited in both digistar and
emacs.

This new implementation of digistar-indent-line-function still needs
work.