digistar-mrslog-mode: further work on syntax highlighting
I am really just making a checkpoint here, as this (simple) way of using
font-lock is not going to cover our needs.  I will rewrite the
highlighter in a later patch, but I just want to tag and release right
here so we can use simple highlighting until that rewrite is done.