require emacs 25.1 or later, instead of 29.1
In 29.1, point-at-bol and point-at-eol are considered obsolete in favor
of pos-bol and pos-eol (or line-beginning-position & line-end-position).
For digistar-mode, which hardly anybody uses, compatibility with more
versions of Emacs is a slight convenience, and the way fields are
handled by these procedures is not relevant, so we will use the older
forms for a while, and maybe update it in the future when we have other
reasons for needing a newer Emacs.

Also an instance of string-replace was able to be changed to
subst-char-in-string, which is more fit for purpose but also compatible
with more versions of Emacs.